requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic-settings>=2.0.0

